+++ PeripheralTech.Model/StaffReview.cs
@@ -12,7 +12,8 @@ namespace PeripheralTech.Model
         public int UserID { get; set; }
         public User User { get; set; }
         public DateTime Date { get; set; }
-        public int Grade { get; set; }
+        //public int Grade { get; set; }
+        public decimal Grade { get; set; }
         public string Specifications { get; set; }
         public string ReviewContent { get; set; }
     }
