+++ PeripheralTech.WebAPI/Database/StaffReview.cs
@@ -13,7 +13,8 @@ namespace PeripheralTech.WebAPI.Database
         public int UserID { get; set; }
         public User User { get; set; }
         public DateTime Date { get; set; }
-        public int Grade { get; set; }
+        //public int Grade { get; set; }
+        public decimal Grade { get; set; }
         public string Specifications { get; set; }
         public string ReviewContent { get; set; }
     }
