+++ PeripheralTech.WebAPI/Database/UserReview.cs
@@ -13,7 +13,8 @@ namespace PeripheralTech.WebAPI.Database
         public int UserID { get; set; }
         public User User { get; set; }
         public DateTime Date { get; set; }
-        public int Grade { get; set; }
+        //public int Grade { get; set; }
+        public decimal Grade { get; set; }
         public string? Comment { get; set; }
     }
 }
